fixing up codegen based on tests on LUMI

+++ ops/c/src/mpi/ops_mpi_hdf5.cpp
@@ -2816,7 +2816,7 @@ void ops_write_plane_hdf5(const ops_dat dat, const int cross_section_dir,
         // TODO if the plane is out of global range, computer range will
         // generate error
         determine_local_range(dat, global_range, local_range);
-        size_t local_buf_size{dat->elem_size};
+        size_t local_buf_size{(size_t)dat->elem_size};
         for (int i = 0; i < space_dim; i++) {
           local_buf_size *= (local_range[2 * i + 1] - local_range[2 * i]);
         }
@@ -2860,7 +2860,7 @@ void ops_write_data_slab_hdf5(const ops_dat dat, const int *range,
     // TODO if the plane is out of global range, computer range will generate
     // error
     determine_local_range(dat, range, local_range);
-    size_t local_buf_size{dat->elem_size};
+    size_t local_buf_size{(size_t)dat->elem_size};
     for (int i = 0; i < space_dim; i++) {
       local_buf_size *= (local_range[2 * i + 1] - local_range[2 * i]);
     }
